make test applicable only to >= go1.10
