+++ credentials/credentials_test.go
@@ -21,9 +21,7 @@ package credentials
 import (
 	"context"
 	"crypto/tls"
-	"crypto/x509"
 	"net"
-	"net/url"
 	"reflect"
 	"strings"
 	"testing"
@@ -360,127 +358,6 @@ func (s) TestAppendH2ToNextProtos(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := appendH2ToNextProtos(tt.ps); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("appendH2ToNextProtos() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func (s) TestParseSpiffeID(t *testing.T) {
-	tests := []struct {
-		name string
-		urls []*url.URL
-		// If we expect ParseSpiffeID to return an error.
-		expectError bool
-		// If we expect TLSInfo.SpiffeID to be plumbed.
-		expectID bool
-	}{
-		{
-			name:        "empty URIs",
-			urls:        []*url.URL{},
-			expectError: false,
-			expectID:    false,
-		},
-		{
-			name: "good SPIFFE ID",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    "foo.bar.com",
-					Path:    "workload/wl1",
-					RawPath: "workload/wl1",
-				},
-				{
-					Scheme:  "https",
-					Host:    "foo.bar.com",
-					Path:    "workload/wl1",
-					RawPath: "workload/wl1",
-				},
-			},
-			expectError: false,
-			expectID:    true,
-		},
-		{
-			name: "invalid host",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    "",
-					Path:    "workload/wl1",
-					RawPath: "workload/wl1",
-				},
-			},
-			expectError: true,
-			expectID:    false,
-		},
-		{
-			name: "invalid path",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    "foo.bar.com",
-					Path:    "",
-					RawPath: "",
-				},
-			},
-			expectError: true,
-			expectID:    false,
-		},
-		{
-			name: "large path",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    "foo.bar.com",
-					Path:    string(make([]byte, 2050)),
-					RawPath: string(make([]byte, 2050)),
-				},
-			},
-			expectError: true,
-			expectID:    false,
-		},
-		{
-			name: "large host",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    string(make([]byte, 256)),
-					Path:    "workload/wl1",
-					RawPath: "workload/wl1",
-				},
-			},
-			expectError: true,
-			expectID:    false,
-		},
-		{
-			name: "multiple SPIFFE IDs",
-			urls: []*url.URL{
-				{
-					Scheme:  "spiffe",
-					Host:    "foo.bar.com",
-					Path:    "workload/wl1",
-					RawPath: "workload/wl1",
-				},
-				{
-					Scheme:  "spiffe",
-					Host:    "bar.baz.com",
-					Path:    "workload/wl2",
-					RawPath: "workload/wl2",
-				},
-			},
-			expectError: false,
-			expectID:    false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			info := TLSInfo{
-				State: tls.ConnectionState{PeerCertificates: []*x509.Certificate{{URIs: tt.urls}}}}
-			err := info.ParseSpiffeID()
-			if got, want := err != nil, tt.expectError; got != want {
-				t.Errorf("want expectError = %v, but got expectError = %v, with error %v", want, got, err)
-			}
-			if got, want := info.SpiffeID != nil, tt.expectID; got != want {
-				t.Errorf("want expectID = %v, but spiffe ID is %v", want, info.SpiffeID)
 			}
 		})
 	}
